Support check_mode in katello_lifecycle_environment

+++ modules/katello_lifecycle_environment.py
@@ -139,14 +139,17 @@ def lifecycle_environment(module, name, organization, state, label=None, descrip
         if current_environment is not None:
             (needs_update, le) = update_fields(desired_environment, current_environment, fields)
             if needs_update:
-                le.update(fields)
+                if not module.check_mode():
+                    le.update(fields)
                 changed = True
         else:
             desired_environment.prior = find_prior(module, "Library", organization) if prior is None else prior
-            desired_environment.create()
+            if not module.check_mode():
+                desired_environment.create()
             changed = True
     elif current_environment is not None:
-        current_environment.delete()
+        if not module.check_mode():
+            current_environment.delete()
         changed = True
     return changed
 
@@ -164,7 +167,8 @@ def main():
             prior=dict(),
             organization=dict(required=True),
             state=dict(required=True),
-        )
+        ),
+        supports_check_mode=True,
     )
 
     if not HAS_NAILGUN_PACKAGE:
